Fixes for half precision

+++ egs/librispeech/ASR/pruned_transducer_stateless7/scaling.py
@@ -423,9 +423,9 @@ class LinearWithAuxLossFunction(torch.autograd.Function):
                 weight_aux_grad = weight.grad
                 alpha_grad = alpha.grad
 
-        x_grad = torch.matmul(ans_grad, weight)
+        x_grad = torch.matmul(ans_grad, weight.to(ans_grad.dtype))
         weight_grad = torch.matmul(ans_grad.reshape(-1, ans_grad.shape[-1]).t(),
-                                   x.reshape(-1, x.shape[-1]))
+                                   x.reshape(-1, x.shape[-1]).to(ans_grad.dtype))
 
         with torch.cuda.amp.autocast(enabled=False):
             weight_grad_norm = weight_grad.to(torch.float32).norm()
